Fixed module containers being unable to get ...
... the composite container as a parent.

Modules should be able to create containers that reference the composite container as a parent.
This can be achieved by using a pre-configured factory. The factory, however, must be able to obtain the
composite container as well.

Classes that use `ModularModuleTrait` should now implement `_getInitialContainer()` in such a way so as to
use the container argument given to set up the container factory service. When modules use this factory to
create their containers, the composite container is assigned as a parent by the factory.